Moved Character to routes
