+++ .env.dev
@@ -1,2 +1,2 @@
-FLASK_APP=web/wsgi.py
+FLASK_APP=wsgi.py
 FLASK_ENV=development
